chore(connect): split FW hash check files
