+++ packages/connect/src/device/Device.ts
@@ -17,7 +17,7 @@ import { DeviceCurrentSession, TypedCallProvider } from './DeviceCurrentSession'
 import { IStateStorage } from './StateStorage';
 import { checkFirmwareRevision } from './checkFirmwareRevision';
 import { getThpChannel } from './thp';
-import { checkFirmwareHashWithRetries } from './workflow/checkFirmwareHash';
+import { checkFirmwareHashWithRetries } from './workflow/checkFirmwareHashWithRetries';
 import { getAllNetworks } from '../data/coinInfo';
 import { getFirmwareStatus, getRelease, getReleaseInfo } from '../data/firmwareInfo';
 import { getLanguage } from '../data/getLanguage';
